rag/requirements.txt
faiss-cpu
numpy
sentence-transformers
python-dotenv
openai
